requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
dj-database-url==0.5.0
Django==3.2
django-config==0.1.4
et-xmlfile==1.1.0
gunicorn==20.1.0
heroku==0.1.4
idna==2.10
kiwisolver==1.3.1
matplotlib==3.4.1
numpy==1.20.2
openpyxl==3.0.7
pandas==1.2.4
Pillow==8.2.0
psycopg2==2.8.6
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.4
python-dotenv==0.17.0
pytz==2021.1
requests==2.25.1
scipy==1.6.3
six==1.15.0
sockets==1.0.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.4
whitenoise==5.2.0
